Validar PU_Nuevo cuando la OC será sin/con Contrato

+++ UanlSISM/UanlSISM/Models/P_OrdenCompra.cs
@@ -10,6 +10,6 @@ namespace UanlSISM.Models
         //public short CANTIDAD { get; set; }
         public short CANTIDAD_NUEVA { get; set; }
         //public short PREUNIT { get; set; }
-        public short PREUNIT_NUEVA { get; set; }
+        public float PREUNIT_NUEVA { get; set; }
     }
 }
